backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
icalendar>=5.0.0
requests>=2.31.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
